Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main() {
-    int a,b,c;
+    float a,b,c;
     printf("");
-    scanf("%d %d %d",&a,&b,&c);
-    printf("Average: %.2d",(a+b+c)/3);
+    scanf("%f %f %f",&a,&b,&c);
+    printf("Average: %.2f",(a+b+c)/3);
     return 0;
 }
